+++ PengurusCRM/src/main/java/com/pengurus/crm/client/service/FileServiceAsync.java
@@ -5,6 +5,6 @@ import java.util.List;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 
 public interface FileServiceAsync {
-	public void getFileList(int quoteId, int jobId, int taskId, int stateId, AsyncCallback<List<String>> callback);
-	public void deleteFile(int quoteId, int jobId, int taskId, int stateId, String fileName, AsyncCallback<Void> callback);
+	public void getFileList(Long quoteId, Long jobId, Long taskId, Long stateId, AsyncCallback<List<String>> callback);
+	public void deleteFile(Long quoteId, Long jobId, Long taskId, Long stateId, String fileName, AsyncCallback<Void> callback);
 }
